added auth middleware in userplaylist

+++ services/playlist/src/routes/userPlaylist.js
@@ -1,34 +1,24 @@
 import express from "express";
 import UserPlaylist from "../models/UserPlaylist";
 import axios from "axios";
-import {
-	uniq
-} from "lodash";
-import {
-	config
-} from "../config";
+import { uniq } from "lodash";
+import { config } from "../config";
+import auth from "../permissions/auth";
 
 const router = express.Router();
-
 
 //Set Base Url
 const baseUrl = `${config.isDev ? config.baseurl.dev : config.baseurl.prod}`;
 
 // create empty Playlist
-router.post("/create", async (req, res) => {
+router.post("/create", auth, async (req, res) => {
 	try {
-		const {
-			name,
-			userId
-		} = req.body;
-
+		const { name } = req.body;
 		const userPlaylist = new UserPlaylist({
 			name,
-			createdBy: userId,
+			createdBy: req.user.id,
 		});
-
 		const playlistData = await userPlaylist.save();
-
 		res.send({
 			status: true,
 			data: playlistData,
@@ -42,33 +32,30 @@ router.post("/create", async (req, res) => {
 });
 
 // add songs into Playlist
-router.post("/addsongs", async (req, res) => {
+router.post("/addsongs", auth, async (req, res) => {
 	try {
-		const {
-			songs,
-			playlistId
-		} = req.body;
-
+		const { songs, playlistId } = req.body;
 		const playlist = await UserPlaylist.findOne({
 			_id: playlistId,
+			createdBy: req.user.id,
 		});
-
-		// yet to hit obs-core addsongs endpoint
+		if (!playlist) {
+			throw new Error("Playlist can't be found...");
+		}
 		const addSongsCoreUrl = `${baseUrl}/addsongs`;
-
 		if (playlist && songs.length) {
 			const availableSongs = playlist.songs;
-			const songIds = songs.map(song => song.videoId)
+			const songIds = songs.map(song => song.videoId);
 			let newSongsList = [...songIds, ...availableSongs];
 			newSongsList = uniq(newSongsList);
 			axios.post(addSongsCoreUrl, {
-				songs: [...songs]
-			})
+				songs: [...songs],
+			});
 			const updateObj = {
 				updatedAt: Date.now(),
 				totalSongs: newSongsList.length,
 				thumbnail: songs[0].thumbnail,
-				songs: newSongsList
+				songs: newSongsList,
 			};
 			await playlist.updateOne(updateObj);
 			res.send({
@@ -76,9 +63,8 @@ router.post("/addsongs", async (req, res) => {
 				data: "songs successfully added!",
 			});
 		} else {
-			throw new Error("something went wrong!")
+			throw new Error("something went wrong!");
 		}
-
 	} catch (error) {
 		res.send({
 			status: false,
@@ -88,18 +74,19 @@ router.post("/addsongs", async (req, res) => {
 });
 
 // get all playlist metadata
-router.get("/getallplaylistmetadata/:uid", async (req, res) => {
+router.get("/getallplaylistmetadata", auth, async (req, res) => {
 	try {
-		const uid = req.params.uid;
-
-		const metaData = await UserPlaylist.find({
-			createdBy: uid,
-		}, {
-			_id: true,
-			name: 1,
-			thumbnail: 2,
-			totalSongs: 3,
-		}, );
+		const metaData = await UserPlaylist.find(
+			{
+				createdBy: req.user.id,
+			},
+			{
+				_id: true,
+				name: 1,
+				thumbnail: 2,
+				totalSongs: 3,
+			}
+		);
 		res.send({
 			status: true,
 			data: metaData,
@@ -113,18 +100,17 @@ router.get("/getallplaylistmetadata/:uid", async (req, res) => {
 });
 
 // get user playlist -  yet to implement song playlist
-router.get("/getplaylist/:id", async (req, res) => {
+router.get("/getplaylist/:id", auth, async (req, res) => {
 	try {
 		const playlistId = req.params.id;
 		const data = await UserPlaylist.findOne({
 			_id: playlistId,
+			createdBy: req.user.id,
 		}).populate("songsList");
-
 		let fetchedAlbum = {
 			...data["_doc"],
-			songs: data["$$populatedVirtuals"]["songsList"]
+			songs: data["$$populatedVirtuals"]["songsList"],
 		};
-
 		res.send({
 			status: true,
 			data: fetchedAlbum,
@@ -132,36 +118,36 @@ router.get("/getplaylist/:id", async (req, res) => {
 	} catch (error) {
 		res.send({
 			status: false,
-			data: error.message
+			data: error.message,
 		});
 	}
 });
 
 // delete songs from playlist
-router.post("/deletesong", async (req, res) => {
+router.post("/deletesong", auth, async (req, res) => {
 	try {
-		const {
-			playlistId,
-			songId
-		} = req.body;
-
-		await UserPlaylist.findByIdAndUpdate(playlistId, {
-			$pull: {
-				songs: songId,
+		const { playlistId, songId } = req.body;
+		await UserPlaylist.findOneAndUpdate(
+			{
+				_id: playlistId,
+				createdBy: req.user.id,
 			},
-		});
-
+			{
+				$pull: {
+					songs: songId,
+				},
+			}
+		);
 		const playlist = await UserPlaylist.findOne({
 			_id: playlistId,
 		});
-
 		const lastSongId = playlist.songs.length ? playlist.songs[playlist.songs.length - 1] : null;
 		let newThumbnail = `https://openbeats.live/static/media/dummy_music_holder.a3d0de2e.jpg`;
 		if (lastSongId !== null) {
 			const songsCoreUrl = `${baseUrl}/getsong/${lastSongId}`;
 			const newThumbData = (await axios.get(songsCoreUrl)).data;
 			if (newThumbData.status) {
-				newThumbnail = newThumbData.data.thumbnail
+				newThumbnail = newThumbData.data.thumbnail;
 			}
 		}
 		await playlist.updateOne({
@@ -183,17 +169,16 @@ router.post("/deletesong", async (req, res) => {
 });
 
 // update name of the playlist
-router.post("/updatename", async (req, res) => {
+router.post("/updatename", auth, async (req, res) => {
 	try {
-		const {
-			name,
-			playlistId
-		} = req.body;
-
+		const { name, playlistId } = req.body;
 		const playlist = await UserPlaylist.findOne({
 			_id: playlistId,
+			createdBy: req.user.id,
 		});
-
+		if (!playlist) {
+			throw new Error("Playlist can't be found...");
+		}
 		await playlist.updateOne({
 			name,
 			updatedAt: Date.now(),
@@ -212,16 +197,17 @@ router.post("/updatename", async (req, res) => {
 });
 
 //delete playlist
-router.get("/delete/:id", async (req, res) => {
+router.get("/delete/:id", auth, async (req, res) => {
 	try {
-		const _id = req.params.id;
-
+		const playlistId = req.params.id;
 		const playlist = await UserPlaylist.findOne({
-			_id,
+			_id: playlistId,
+			createdBy: req.user.id,
 		});
-
+		if (!playlist) {
+			throw new Error("Playlist can't be found...");
+		}
 		await playlist.deleteOne();
-
 		res.send({
 			status: true,
 			data: "playlist deleted successfully!",
